requirements.txt
psutil
Flask
mysql-connector-python
Flask-Cors